render katex in a hugo shortcode; use a shared intersectionobserver
This is one step toward being able to do static rendering of KaTeX.  The
downside is that we either have to write very verbose Hugo shortcode
tags when we type in math, or we have to do a regexp pass. If we do a
regexp pass, unfortunately *this* will have an impact on embedded LaTeX
code that is meant to be rendered externally; therefore, it is necessary
for the concrete syntax of KaTeX embedding to not clash with that of
math mode in LaTeX. For this purpose, I reserve {#...#}.

Finally I have updated the rendering code to use a shared IntersectionObserver for
visibility rather than dozens of little ones. This is claimed to be
significantly better performing: https://blog.webdevsimplified.com/2022-01/intersection-observer/