huge breaking change: strict import semantics
Imports were silently transitive in the old version which is a terrible
idea; moreover, the meaning of expanding a macro that calls another macro
is kind of unclear when that inner macro is defined in the importing
module... What a mess.

I have now changed Forester to have standard non-transitive import
semantics, and macro expansion is really just functional reduction now.
In other words, the bodies of macros are resolved at definition-time
rather than at use-time, just like functions.